apps/requirements.txt
googletrans==4.0.0rc1
asyncio
mysql-connector-python